% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_occ.R
\name{get_gbif_points}
\alias{get_gbif_points}
\title{Search for GBIF occurrences}
\usage{
get_gbif_points(key, gbif_limit, out_format)
}
\arguments{
\item{key}{A GBIF name/taxon key}

\item{gbif_limit}{Number of records to return}

\item{out_format}{Format of output file options = gbif(Darwin Core), IUCN SIS, combined (SIS and gbif)}
}
\value{
Returns a data frame with occurrence points
}
\description{
Search for GBIF occurrences
}
\examples{

}
\keyword{GBIF,}
\keyword{KNMS,}
\keyword{Online}
\keyword{Plants}
\keyword{World}
\keyword{of}
\keyword{the}
