% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCRapid-package.R
\docType{package}
\name{LCRapid-package}
\alias{LCRapid}
\alias{LCRapid-package}
\title{LCRapid: Generate a rapid Least Concern Red List assessment for plant species}
\description{
Quickly determine if a plant species is likely to be non-threatened (Least Concern - See IUCN Red List). Generate minimal documentation for a Least Concern species and submit to IUCN Red List via https://connect.iucnredlist.org/ (registration needed)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stevenpbachman/LCRapid}
  \item Report bugs at \url{https://github.com/stevenpbachman/LCRapid/issues}
}

}
\author{
\strong{Maintainer}: Steven Bachman \email{stevenpbachman@gmail.com} (\href{https://orcid.org/0000-0003-1085-6075}{ORCID})

}
\keyword{internal}
