% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_search.R
\name{name_search}
\alias{name_search}
\title{Search plant name against GBIF, KNMS and POWO names backbone}
\usage{
name_search(name)
}
\arguments{
\item{name}{A scientific plant species name. Better results can be obtained
when author is included e.g. Poa annua L.}
}
\value{
Returns a data frame with ...
}
\description{
Perform a fuzzy search against three names list to get best match and POWO taxonomic status
}
\examples{
# single name search
name_search("Poa annua L.")

# Or, search multiple names using purrr::map_dfr
names <- c("Poa annua L.", "Welwitschia mirabilis Hook.f.")
names_out <- purrr::map_dfr(names, name_search)
}
\keyword{GBIF,}
\keyword{KNMS,}
\keyword{Online}
\keyword{Plants}
\keyword{World}
\keyword{of}
\keyword{the}
