% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_search_gbif.R
\name{name_search_gbif}
\alias{name_search_gbif}
\title{Search plant name against GBIF names backbone}
\usage{
name_search_gbif(name)
}
\arguments{
\item{name}{A plant species binomial with or without author e.g. Poa annua L.}
}
\value{
Returns a data frame with IPNI identifier, binomial, author and accepted status
}
\description{
This function uses rgbif package to search a name against the GBIF names backbone.
}
\examples{
name_search_gbif("Poa annua L.")
}
\keyword{GBIF}
