% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_name_powo.R
\name{search_name_powo}
\alias{search_name_powo}
\title{Search plant name in POWO}
\usage{
search_name_powo(name_in)
}
\arguments{
\item{name_in}{A plant species binomial e.g. Poa annua}
}
\value{
Returns a data frame with IPNI identifier, binomial, author and accepted status
}
\description{
This function searches Plants of the World Online (http://www.plantsoftheworldonline.org/) and returns name status
}
\examples{
search_name_powo('Poa annua')
}
\keyword{POWO}
