% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_search_gbif.R
\name{name_search_gbif}
\alias{name_search_gbif}
\title{Search plant name against GBIF names backbone}
\usage{
name_search_gbif(name)
}
\arguments{
\item{name}{A scientific plant species name. Better results can be obtained
when author is included e.g. Poa annua L.}
}
\value{
Returns a data frame with initial search term \code{searchName}, GBIF taxon key \code{usageKey}, GBIF scientific
name \code{scientificName}, and a measure of \code{confidence} in the match. When there is no match it returns a value
of "no_match" under the \code{confidence} field.
}
\description{
This function uses the \code{rgbif} package \url{https://docs.ropensci.org/rgbif/index.html}
to query a scientific plant name against the GBIF names backbone - see \url{https://www.gbif.org/}.
It restricts the result to species rank constrained to kingdom = 'Plantae'.
This function uses the \code{rgbif::name_backbone_verbose} function.
}
\examples{
# single name search
name_search_gbif("Poa annua L.")

# Or, search multiple names using purrr::map_dfr
names <- c("Poa annua L.", "Welwitschia mirabilis Hook.f.")

#names_out <- purrr::map_dfr(names, name_search_gbif)
}
\keyword{GBIF}
